Little simple refactoring: 1. Enabled the usage of more safe alternatives to malloc and strdup; 2. Improved debug messages; 3. Fixed some memory leaks with usage of cleanup 4. Make comments more clear and simple to understand 5. Removed elapsed time output (added it before for comparing with old realization) 6. Improved code-style a little
