+++ data-structures/lazy-segment-tree.h
@@ -2,6 +2,7 @@
 #define LAZY_SEGMENT_TREE
 
 #include "stddef.h"
+#include "../utils.h"
 
 /// --------------------------------------------------------------------------------------------------------------------
 /// Type declarations
